refactor(bot-client): rename contextWritePath → gatewayWriteHelpers
Post-CONTEXT_MODE cleanup: beta.135 removed the mode-switching flags
(isRawEnvelopeEnabled/isThinPayloadEnabled), leaving the module as exactly three
gateway write helpers (persistUserMessageViaGateway, persistAssistantMessageViaGateway,
syncConversationViaGateway). The old name described the deleted mode-switching
"write path"; gatewayWriteHelpers names what it is now and joins the existing
gateway* utility family (gatewayServiceCalls, gatewayClients, gatewayFetchGuard).

Pure rename — file + colocated test moved via git mv, logger label and the two
production importers (ConversationPersistence, SyncExecutor) plus their test
imports/mocks repointed. No behaviour change.
